put schema under deepar
